LayoutTests/imported/w3c/web-platform-tests/css/css-values/minmax-time-serialize-expected.txt


FAIL 'min(1ms)' as a specified value should serialize as 'calc(0.001s)'. assert_equals: 'calc(0.001s)' should round-trip exactly in specified values. expected "calc(0.001s)" but got "calc(1ms)"
PASS 'min(1ms)' as a computed value should serialize as '0.001s'. 
FAIL 'min(1s)' as a specified value should serialize as 'calc(1s)'. assert_equals: 'calc(1s)' should round-trip exactly in specified values. expected "calc(1s)" but got "calc(1000ms)"
PASS 'min(1s)' as a computed value should serialize as '1s'. 
FAIL 'max(1ms)' as a specified value should serialize as 'calc(0.001s)'. assert_equals: 'calc(0.001s)' should round-trip exactly in specified values. expected "calc(0.001s)" but got "calc(1ms)"
PASS 'max(1ms)' as a computed value should serialize as '0.001s'. 
FAIL 'max(1s)' as a specified value should serialize as 'calc(1s)'. assert_equals: 'calc(1s)' should round-trip exactly in specified values. expected "calc(1s)" but got "calc(1000ms)"
PASS 'max(1s)' as a computed value should serialize as '1s'. 
FAIL 'min(1s, 2s, 3s)' as a specified value should serialize as 'calc(1s)'. assert_equals: 'calc(1s)' should round-trip exactly in specified values. expected "calc(1s)" but got "calc(1000ms)"
PASS 'min(1s, 2s, 3s)' as a computed value should serialize as '1s'. 
FAIL 'min(3s, 2s, 1s)' as a specified value should serialize as 'calc(1s)'. assert_equals: 'calc(1s)' should round-trip exactly in specified values. expected "calc(1s)" but got "calc(1000ms)"
PASS 'min(3s, 2s, 1s)' as a computed value should serialize as '1s'. 
FAIL 'max(1s, 2s, 3s)' as a specified value should serialize as 'calc(3s)'. assert_equals: 'calc(3s)' should round-trip exactly in specified values. expected "calc(3s)" but got "calc(3000ms)"
PASS 'max(1s, 2s, 3s)' as a computed value should serialize as '3s'. 
FAIL 'max(3s, 2s, 1s)' as a specified value should serialize as 'calc(3s)'. assert_equals: 'calc(3s)' should round-trip exactly in specified values. expected "calc(3s)" but got "calc(3000ms)"
PASS 'max(3s, 2s, 1s)' as a computed value should serialize as '3s'. 
FAIL 'min(900ms, 1s)' as a specified value should serialize as 'calc(0.9s)'. assert_equals: 'calc(0.9s)' should round-trip exactly in specified values. expected "calc(0.9s)" but got "calc(900ms)"
PASS 'min(900ms, 1s)' as a computed value should serialize as '0.9s'. 
FAIL 'max(1100ms, 1s)' as a specified value should serialize as 'calc(1.1s)'. assert_equals: 'calc(1.1s)' should round-trip exactly in specified values. expected "calc(1.1s)" but got "calc(1100ms)"
PASS 'max(1100ms, 1s)' as a computed value should serialize as '1.1s'. 
FAIL 'calc(min(1s, 2s) + max(3s, 4s) + 10s)' as a specified value should serialize as 'calc(15s)'. assert_equals: 'calc(15s)' should round-trip exactly in specified values. expected "calc(15s)" but got "calc(15000ms)"
PASS 'calc(min(1s, 2s) + max(3s, 4s) + 10s)' as a computed value should serialize as '15s'. 
